proradilo je dodavanje setova

+++ app/src/main/java/com/example/myapplication/data/auth/requests/SetRequest.kt
@@ -1,9 +1,9 @@
 package com.example.myapplication.data.auth.requests
 
 data class SetRequest (
-    val sessionExerciseId: Long,
-    val setNumber: Integer,
-    val weight: Integer,
-    val reps: Integer,
-    val rpe: Integer
+    val sessionExerciseId: Int,
+    val setNumber: Int,
+    val weight: Double,
+    val reps: Int,
+    val rpe: Int
 )
